Streamlining code.
send_simple_products is now sending products to WC with price and category
but not SKU nor Images.  SKU was working intermittently an hour ago.

woo_product appears to have a lot of extra code.  Going to refactor out
anything that isn't currently being used.  It can be put back in later.